pyproject.toml
[tool.poetry]
name = "health-sms"
version = "0.1.0"
description = "SMS powered tool for storing and retrieving your health-based data."
authors = ["wookiee89 <jay.r.jain@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "health_sms"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
pydantic-settings = "^2.1.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
